Refactor: move api file path
