--- src/pages/my-fights/api/delete-my-fights-data.tsx
@@ -1,10 +1,0 @@
-import { doc, getFirestore, deleteDoc } from "firebase/firestore";
-import { app } from "src/shared/FireBase";
-
-const db = getFirestore(app);
-
-const deleteMyFightsData = (id: string) => {
-  return deleteDoc(doc(db, "myFights", id));
-};
-
-export default deleteMyFightsData;
